net, netmap: implement security check on size of specified vma
The linux_netmap_mmap() hook does not include any check on the size
of the virtual memory area specified via the mmap() system call. Any
value smaller than the actual total size of the netmap area structures
would cause a failure. This patch implements a security check on the
specified vma size and returns -EINVAL in case of mismatch to report
the wrong value correctly to the user application.

Signed-off-by: Stefano Alletto <alletto.s@gmail.com>
Signed-off-by: Arianna Avanzini <avanzini.arianna@gmail.com>
Signed-off-by: Daniele Cristofori <daniele.cristofori89@gmail.com>